Remove key "xml" from the example
I'm not sure what this is useful for, the docs for this mention that you
can use swagger yaml and xml interchangibly so I suppose some parsers,
maybe even a tool by swagger/openapi itself exists that does this but I
haven't checked. This library certainly so far does not do anything with
this key, nor does it have xml support, I'm therefore removing it for now.

https://swagger.io/docs/specification/data-models/representing-xml/